% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullData.R
\name{download_ana_flow}
\alias{download_ana_flow}
\title{Download ANA flow data for stations}
\usage{
download_ana_flow(station_codes, start_date, end_date)
}
\arguments{
\item{station_codes}{vector of station IDs}

\item{start_date}{character "dd/mm/yyyy"}

\item{end_date}{character "dd/mm/yyyy"}

\item{out_dir}{directory to save CSV files}
}
\value{
NULL (writes CSV files)
}
\description{
Download ANA flow data for stations
}
