% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullData.R
\name{get_ana_metadata}
\alias{get_ana_metadata}
\title{Get ANA station metadata (cached locally)}
\usage{
get_ana_metadata()
}
\arguments{
\item{path}{optional custom file path}
}
\value{
sf object
}
\description{
Get ANA station metadata (cached locally)
}
