% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullData.R
\name{parse_ana_month}
\alias{parse_ana_month}
\title{Reformat monthly ANA data to readable format}
\usage{
parse_ana_month(month_xml, station, date_raw)
}
\arguments{
\item{date_raw}{}
}
\description{
Reformat monthly ANA data to readable format
}
