% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullData.R
\name{select_stations}
\alias{select_stations}
\title{Select ANA stations by basin and spatial intersection}
\usage{
select_stations(meta, basins)
}
\arguments{
\item{meta}{sf object of station metadata}

\item{basins}{character vector of basin names (e.g., c("RIO AMAZONAS"))}

\item{boundary}{sf object of basin boundaries}
}
\value{
sf object of filtered stations
}
\description{
Select ANA stations by basin and spatial intersection
}
